+++ src/app/shared/components/web-footer.ts
@@ -1,10 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 
 import OJP_Legacy from '../../config/ojp-legacy';
-import { DEBUG_LEVEL } from '../../config/constants';
+import * as OJP_Next from 'ojp-sdk-next';
+
+import { DEBUG_LEVEL, OJP_VERSION } from '../../config/constants';
 
 type PageModel = {
-  sdkVersion: string,
+  sdkVersionText: string,
   lastUpdate: string,
   isTopograhicPlaceMapLayerEnabled: boolean,
   changelogURL: string,
@@ -20,20 +22,21 @@ export class WebFooterComponent implements OnInit {
 
   constructor() {
     const isTopograhicPlaceMapLayerEnabled = DEBUG_LEVEL === 'DEBUG';
+    const isOJPv2 = OJP_VERSION === '2.0';
 
     let changelogURL = 'https://github.com/openTdataCH/ojp-demo-app-src/blob/feature/ojp-v1-beta/CHANGELOG.md';
-    if (OJP_Legacy.OJP_VERSION === '2.0') {
+    if (isOJPv2) {
       changelogURL = 'https://github.com/openTdataCH/ojp-demo-app-src/blob/feature/ojp-v2-beta/CHANGELOG.md';
     }
 
     let githubURL = 'https://github.com/openTdataCH/ojp-js/blob/feature/ojp-v1/CHANGELOG.md';
-    if (OJP_Legacy.OJP_VERSION === '2.0') {
+    if (isOJPv2) {
       githubURL = 'https://github.com/openTdataCH/ojp-js/blob/feature/ojp-v2/CHANGELOG.md';
     }
 
     this.model = {
-      sdkVersion: OJP_Legacy.SDK_VERSION + ', OJP API ' + OJP_Legacy.OJP_VERSION,
-      lastUpdate: '4.June 2025',
+      sdkVersionText: 'ojp-sdk-legacy: ' + OJP_Legacy.SDK_VERSION + ' - ojp-sdk-next: ' + OJP_Next.SDK_VERSION + ' - OJP API ' + OJP_VERSION,
+      lastUpdate: '6.June 2025',
       isTopograhicPlaceMapLayerEnabled: isTopograhicPlaceMapLayerEnabled,
       changelogURL: changelogURL,
       githubURL: githubURL,