+++ src/app/shared/models/trip-info-result.ts
@@ -2,6 +2,8 @@ import * as OJP_Types from 'ojp-shared-types';
 import * as OJP_Next from 'ojp-sdk-next';
 import OJP_Legacy from '../../config/ojp-legacy';
 
+import { OJP_VERSION } from '../../config/constants';
+
 import { StopEventType, StopPointCall, StopPointCallType } from '../types/_all';
 import { OJPHelpers } from '../../helpers/ojp-helpers';
 import { JourneyService } from './journey-service';
@@ -136,7 +138,9 @@ export class TripInfoResult {
     });
 
     const journeyService: JourneyService | null = (() => {
-      if (OJP_Legacy.OJP_VERSION === '1.0') {
+      const isOJPv2 = OJP_VERSION === '2.0';
+
+      if (!isOJPv2) {
         const oldTripInfoResultSchema = firstTripInfoResultSchema as OJP_Types.OJPv1_TripInfoResultStructureSchema;
         const service = JourneyService.initWithLegacyTripInfoResultSchema(oldTripInfoResultSchema);
         return service;