+++ src/app/station-board/search/custom-stop-event-xml-popover/custom-stop-event-xml-popover.component.ts
@@ -3,6 +3,8 @@ import { UserTripService } from 'src/app/shared/services/user-trip.service';
 
 import OJP_Legacy from '../../../config/ojp-legacy';
 
+import { REQUESTOR_REF, OJP_VERSION } from '../../../config/constants';
+
 @Component({
   selector: 'custom-stop-event-popover',
   templateUrl: './custom-stop-event-xml-popover.component.html',
@@ -24,10 +26,12 @@ export class CustomStopEventXMLPopoverComponent {
   }
 
   public parseCustomRequestXML() {
-    this.isRunningRequest = true
+    this.isRunningRequest = true;
+    const isOJPv2 = OJP_VERSION === '2.0';
+    const xmlConfig = isOJPv2 ? OJP_Legacy.XML_ConfigOJPv2 : OJP_Legacy.XML_BuilderConfigOJPv1;
 
     const stageConfig = this.userTripService.getStageConfig();
-    const request = OJP_Legacy.StopEventRequest.initWithRequestMock(this.customRequestXMLs, stageConfig);
+    const request = OJP_Legacy.StopEventRequest.initWithRequestMock(this.customRequestXMLs, stageConfig, xmlConfig, REQUESTOR_REF);
     request.fetchResponse().then(response => {
       this.isRunningRequest = false;
       const responseXML = request.requestInfo.responseXML;